Move packages to api folder
