+++ api/controllers/base.go
@@ -11,7 +11,7 @@ import (
 	_ "github.com/jinzhu/gorm/dialects/mysql"    //mysql database driver
 	_ "github.com/jinzhu/gorm/dialects/postgres" //postgres database driver
 
-	"github.com/victorsteven/fullstack/api/models"
+	"github.com/trecky/blog-dashboard-golang-api/api/models"
 )
 
 type Server struct {
