+++ api/controllers/home_controller.go
@@ -3,7 +3,7 @@ package controllers
 import (
 	"net/http"
 
-	"github.com/victorsteven/fullstack/api/responses"
+	"github.com/trecky/blog-dashboard-golang-api/api/responses"
 )
 
 func (server *Server) Home(w http.ResponseWriter, r *http.Request) {
