+++ api/controllers/login_controller.go
@@ -5,10 +5,10 @@ import (
 	"io/ioutil"
 	"net/http"
 
-	"github.com/victorsteven/fullstack/api/auth"
-	"github.com/victorsteven/fullstack/api/models"
-	"github.com/victorsteven/fullstack/api/responses"
-	"github.com/victorsteven/fullstack/api/utils/formaterror"
+	"github.com/trecky/blog-dashboard-golang-api/api/auth"
+	"github.com/trecky/blog-dashboard-golang-api/api/models"
+	"github.com/trecky/blog-dashboard-golang-api/api/responses"
+	"github.com/trecky/blog-dashboard-golang-api/api/utils/formaterror"
 	"golang.org/x/crypto/bcrypt"
 )
