+++ api/controllers/routes.go
@@ -1,6 +1,6 @@
 package controllers
 
-import "github.com/victorsteven/fullstack/api/middlewares"
+import "github.com/trecky/blog-dashboard-golang-api/api/middlewares"
 
 func (s *Server) initializeRoutes() {
