+++ api/middlewares/middlewares.go
@@ -4,8 +4,8 @@ import (
 	"errors"
 	"net/http"
 
-	"github.com/victorsteven/fullstack/api/auth"
-	"github.com/victorsteven/fullstack/api/responses"
+	"github.com/trecky/blog-dashboard-golang-api/api/auth"
+	"github.com/trecky/blog-dashboard-golang-api/api/responses"
 )
 
 func SetMiddlewareJSON(next http.HandlerFunc) http.HandlerFunc {
