+++ go.mod
@@ -5,9 +5,9 @@ go 1.14
 require (
 	github.com/badoux/checkmail v0.0.0-20181210160741-9661bd69e9ad // indirect
 	github.com/dgrijalva/jwt-go v3.2.0+incompatible // indirect
-	github.com/gorilla/mux v1.7.4 // indirect
-	github.com/jinzhu/gorm v1.9.12 // indirect
-	github.com/joho/godotenv v1.3.0 // indirect
-	golang.org/x/crypto v0.0.0-20200510223506-06a226fb4e37 // indirect
+	github.com/gorilla/mux v1.7.4
+	github.com/jinzhu/gorm v1.9.12
+	github.com/joho/godotenv v1.3.0
+	golang.org/x/crypto v0.0.0-20200510223506-06a226fb4e37
 	gopkg.in/go-playground/assert.v1 v1.2.1 // indirect
 )
