+++ main.go
@@ -1,7 +1,6 @@
-
 package main
 
-import "https://github.com/trecky/blog-dashboard-golang-api/api"
+import "github.com/trecky/blog-dashboard-golang-api/api"
 
 func main() {
 	api.Run()
